+++ src/pgen/pgen.hpp
@@ -86,6 +86,13 @@ using namespace parthenon::driver::prelude;
 
 void ProblemGenerator(Mesh *pm, parthenon::ParameterInput *pin, MeshData<Real> *md);
 } // namespace kh
+
+namespace lw_implode {
+using namespace parthenon::driver::prelude;
+
+void ProblemGenerator(MeshBlock *pmb, parthenon::ParameterInput *pin);
+} // namespace lw_implode
+
 namespace rand_blast {
 using namespace parthenon::driver::prelude;