+++ CHANGELOG.md
@@ -10,6 +10,7 @@ In between a subtle bug was introduced that resulted in inconsistent divergence 
 refinement.
 
 ### Added (new features/APIs/variables/...)
+- [[PR 140]](https://github.com/parthenon-hpc-lab/athenapk/pull/140) Add hydro reflecting boundary conditions
 - [[PR 89]](https://github.com/parthenon-hpc-lab/athenapk/pull/89) Add viscosity and resistivity
 - [[PR 1]](https://github.com/parthenon-hpc-lab/athenapk/pull/1) Add isotropic thermal conduction and RKL2 supertimestepping